🐛 fix: 사용자 가이드 도중에 닫는 경우 상태 초기화
