+++ apps/client/src/shared/store/useCoachMarkStore.ts
@@ -12,6 +12,6 @@ export const useCoachMarkStore = create<TCoachMarkStore>()((set) => ({
   isCoachMarkOpen: true,
   currentStep: 0,
   openCoachMark: () => set({ isCoachMarkOpen: true, currentStep: 0 }),
-  closeCoachMark: () => set({ isCoachMarkOpen: false }),
+  closeCoachMark: () => set({ isCoachMarkOpen: false, currentStep: 5 }),
   setCurrentStep: (step) => set({ currentStep: step }),
 }));
