+++ apps/client/src/widgets/workspace/CoachMark/CoachMark.tsx
@@ -1,16 +1,10 @@
-import { useEffect } from 'react';
 import { useCoachMarkStore } from '@/shared/store/useCoachMarkStore';
 import { CircleButton } from '@/shared/ui';
-import * as Blockly from 'blockly/core';
-import TabbedToolbox from '@/core/tabbedToolbox';
 import { coachMarkContent } from '@/shared/utils';
 
 export const CoachMark = () => {
   const { currentStep, setCurrentStep, closeCoachMark } = useCoachMarkStore();
   const stepsLength = coachMarkContent.length;
-  const toolboxDiv = document.querySelector('.blocklyToolboxDiv');
-  // const blockCanvas = document.querySelector('.blocklyBlockCanvas');
-  // TODO: 사용자가이드 - 블록 하이라이팅
 
   const nextStep = () => {
     if (currentStep < stepsLength - 1) {
@@ -28,38 +22,6 @@ export const CoachMark = () => {
     localStorage.setItem('isCoachMarkDismissed', 'true');
     closeCoachMark();
   };
-
-  useEffect(() => {
-    const workspace = Blockly.getMainWorkspace() as Blockly.WorkspaceSvg;
-    if (!workspace) {
-      return;
-    }
-
-    const toolbox = workspace.getToolbox() as TabbedToolbox;
-
-    if (!toolbox) return;
-
-    switch (currentStep) {
-      case 0:
-        toolbox.clickTab('html');
-        break;
-      case 1:
-        toolbox.clickTab('css');
-        break;
-      case 2:
-        toolbox.clickTab('html');
-    }
-
-    if (toolboxDiv) {
-      if (currentStep <= 1) {
-        toolboxDiv.classList.add('coachMarkHighlight');
-        // blockCanvas.classList.add('coachMarkHighlight');
-      } else {
-        toolboxDiv.classList.remove('coachMarkHighlight');
-        // blockCanvas.classList.remove('coachMarkHighlight');
-      }
-    }
-  }, [currentStep, toolboxDiv]);
 
   return (
     <div className="fixed inset-0 z-[100] flex items-center justify-center bg-black bg-opacity-70">
@@ -83,7 +45,6 @@ export const CoachMark = () => {
               그만 보기
             </CircleButton>
 
-            {/* TODO: 이전 버튼 - 기능 구현 후 삭제 */}
             <CircleButton
               className="text-bold-sm h-8 w-16"
               onClick={prevStep}
